+++ lib/capybara_test_helpers/test_helper.rb
@@ -60,18 +60,16 @@ class CapybaraTestHelpers::TestHelper
     raise_missing_element_error
   end
 
-  # Public: Wraps a Capybara::Node::Element with a test helper object.
-  def wrap_element(capybara_node)
-    if capybara_node.is_a?(Enumerable)
-      capybara_node.map { |node| wrap_element(node) }
+  # Public: Wraps a Capybara::Node::Element or test helper with a test helper
+  # object of this class.
+  def wrap_element(element)
+    if element.is_a?(Enumerable)
+      element.map { |node| wrap_element(node) }
     else
-      self.class.new(capybara_node, test_context: test_context)
-    end
-  end
+      raise ArgumentError, "#{ element.inspect } must be a test helper or element." unless element.respond_to?(:to_capybara_node)
 
-  # Public: Wraps a CapybaraTestHelper with a different test helper object.
-  def wrap_test_helper(test_helper)
-    self.class.new(test_helper.query_context, test_context: test_context) if test_helper.query_context
+      self.class.new(element.to_capybara_node, test_context: test_context)
+    end
   end
 
   # Public: Scopes the Capybara queries in the block to be inside the specified
@@ -140,7 +138,7 @@ private
       delegate(*method_names, to: :test_context)
     end
 
-    # Public: Allows to define dependencies on other matchers.
+    # Public: Allows to make other test helpers available.
     #
     # NOTE: When you call a helper the "negated" state is preserved for assertions.
     #
@@ -153,11 +151,7 @@ private
       helper_names.each do |helper_name|
         private define_method(helper_name) { |element = nil|
           test_helper = test_context.get_test_helper(helper_name)
-          if element
-            raise ArgumentError, "#{ element.inspect } must be a test helper or element." unless element.respond_to?(:to_capybara_node)
-
-            test_helper = test_helper.wrap_element(element.to_capybara_node)
-          end
+          test_helper = test_helper.wrap_element(element) if element
           @negated.nil? ? test_helper : test_helper.should(@negated)
         }
       end