+++ spec/support/default_test_helpers.rb
@@ -4,10 +4,10 @@
 module DefaultTestHelpers
   def self.included(base)
     base.instance_eval do
-      use_test_helpers(:current_page, :routes)
+      use_test_helpers(:current_page, :navigation)
 
       # We delegate this method because it's used extremely often (on every test!)
-      delegate :visit_page, to: :routes
+      delegate :visit_page, to: :navigation
 
       # Ensure we use a consistent screen size.
       before(:each, :screen_size) do |example|