+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Capybara Test Helpers 1.0.4 (2020-11-26) ##
+
+* Add `have_no` alias for `have_no_selector`.
+* Add `has_no?` alias for `has_no_selector?`.
+* Remove internal method `wrap_test_helper`.
+* Allow to use locator aliases in `assert_` methods as well, for consistency.
+
 ## Capybara Test Helpers 1.0.3 (2020-11-24) ##
 
 * Add `CapybaraTestHelpers::BenchmarkHelpers`